blog_api/requirements.txt
aiofiles==0.5.0
aioredis==1.3.1
aiosmtplib==1.1.6
alembic==1.6.5
aniso8601==7.0.0
anyio==3.2.1
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==3.0.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.5.2
fastapi==0.65.2
fastapi-mail==0.4.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.0
h11==0.12.0
hiredis==2.0.0
httpcore==0.13.6
httptools==0.1.2
httpx==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==2.0.1
orjson==3.5.3
passlib==1.7.4
promise==2.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
rfc3986==1.5.0
rsa==4.7.2
Rx==1.6.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.18
starlette==0.14.2
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.5
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
